requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
idna==3.3
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
twilio==7.5.1
urllib3==1.26.8
